Change element type in interpreter

+++ Brainfuck.Core/Interpreter.cs
@@ -2,7 +2,7 @@
 
 namespace Brainfuck.Core
 {
-    using BufferElement = Byte;
+    using BufferElement = Int32;
 
     public static class Interpreter
     {
